CafeCrazy
This project Cafe Crazy Cheesy has been developed on HTML and CSS. The main aim of developing this project is s to create a static website for the Cafe, from which user can get the details of the Prize, Menu such as about the Cafe, contact details, address, images etc. This project is in HTML and CSS so it's a Land static website and does not perform any type of dynamic database operations.

Pages of the project Cafe Crazy Cheesy: • Home Page: This project contains the basic information. • About Page: It contains all about the Cafe Crazy Cheesy information and their working. • Menu Page: This page contain all images and information about Menu of the Cafe Crazy Cheesy. • Contact Page: User can submit their contact details on the website. • Login: User can log in on the website.